solving bugs at 08/06/2024

+++ project/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render, get_object_or_404, redirect
-from .forms import ProjectForm, SupplierForm, Supplier_contactForm,ClientForm, RepresentativeForm, BuyerForm
+from .forms import ProjectForm, SupplierForm, Supplier_contactForm,ClientForm, BuyerForm
 from .models import Project, Supplier, Client, Country, Language, File, Representative, Buyer, Representative
 from django.contrib import messages
 from django.http import JsonResponse
@@ -14,65 +14,70 @@ from datetime import datetime
 
 def get_project_nbr():
     P_PREFIX = 'P'
+    last_rank = 0
     latest_project = Project.objects.order_by('-id').first()
     current_year = datetime.now().strftime("%y")
-    if latest_project and latest_project.project_nbr.startswith(f'{P_PREFIX}{current_year}'):
+    if latest_project and latest_project.rank is not None and \
+    latest_project.project_nbr.startswith(f'{P_PREFIX}{current_year}'):
         last_rank = latest_project.rank + 1
         project_nbr = f'{P_PREFIX}{current_year}{last_rank:04d}'
     else:
-        project_nbr = f'{P_PREFIX}{current_year}0000'
-    return project_nbr
+        project_nbr = f'{P_PREFIX}{current_year}{last_rank:04d}'
+    return last_rank, project_nbr
 
 @login_required(login_url='sign-in')
 def project_home(request):
-    clients = Client.objects.all()
-    project_nbr  = get_project_nbr()
+    _, project_nbr  = get_project_nbr()
     representatives = Representative.objects.all()
-    buyers = Buyer.objects.all()
-    context = {'clients':clients,
-               'representatives': representatives,
-               'buyers': buyers,
+    context = {'representatives': representatives,
                'project_nbr':project_nbr}
     return render(request, 'project_home.html',context)
 
 
-def project_detail(request, project_nbr):
-    project = get_object_or_404(Project, project_nbr=project_nbr)
-    clients = Client.objects.all()
-    articles = project.article_set.all()[:50]
-    Representatives = Representative.objects.all()
-    context = {'project':project,
-               'clients':clients,
-               'page':'update-project', 
-               'representatives':Representatives,
-               'articles':articles}
+def project_detail(request, pk):
+    try:
+        project = Project.objects.get(id=pk)
+        Representatives = Representative.objects.all()
+        context = {'project':project,
+                   'representatives':Representatives}
+    except Project.DoesNotExist:
+        messages.error(request, 'Project not found !')
     return render(request, 'project_home.html', context)
 
-def project_create(request):
+def get_message_error(form):
+    errors = form.errors.as_data()
+    custom_errors = ""
+    for field, field_errors in errors.items():
+        for error in field_errors:
+            custom_errors += f"Error in field '{field}': {error.message}\n"
+    return custom_errors
+
+def create_project(request):
     if request.method == 'POST':
         form = ProjectForm(request.POST)
         if form.is_valid():
-            project = form.save()
+            project = form.save(commit=False)
+            last_rank,_  = get_project_nbr()
+            project.rank = last_rank
+            project.save()
             messages.success(request, 'Project has been created successfully')
-            return project_detail(request,project.project_nbr)
-        for err in form.errors:
-            print(err)
-        messages.error(request, 'An error occured, please retry')
+            return redirect('project-detail', project.id)
+        messages.error(request, get_message_error(form))
     return redirect('project-home')
 
-def project_edit(request, pk):
+def update_project(request, pk):
     project = get_object_or_404(Project, pk=pk)
     if request.method == 'POST':
         form = ProjectForm(request.POST, instance=project)
         if form.is_valid():
             project = form.save()
             messages.success(request, 'Project has been updated successfully')
-        for err in form.errors:
-            print(err)
-    return redirect('project-detail', project.project_nbr)
+        else:
+            messages.error(request, get_message_error(form))
+    return redirect('project-detail', project.id)
 
 
-def project_delete(request, pk):
+def delete_project(request, pk):
     if request.method == 'POST':
         project = get_object_or_404(Project, pk=pk)
         project.delete()
@@ -81,17 +86,21 @@ def project_delete(request, pk):
 
 def add_article_to_project(request):
     if request.method == 'POST':
-        project_nbr = request.POST['project_nbr']
-        article_nbr = request.POST['article_nbr']
-        if project_nbr and article_nbr:
-            try:
-                project = Project.objects.get(project_nbr=project_nbr)
-                article = Article.objects.get(article_nbr=article_nbr)
-                article.projects.add(project)
-                messages.success(request,'Article has been added to project succesfully')
-            except:
-                messages.error(request,'An error occured ! please retry again')
-        return redirect('project-detail', project_nbr)
+        project_nbr = request.POST.get('project_nbr')
+        article_nbr = request.POST.get('article_nbr')
+        try:
+            project = Project.objects.get(project_nbr=project_nbr)
+            article = Article.objects.get(article_nbr=article_nbr)
+            article.projects.add(project)
+            messages.success(request, 'Article has been added to project successfully.')
+            return redirect('project-detail', project.id)
+        except Project.DoesNotExist:
+            messages.error(request, f"Project with number '{project_nbr}' not found.")
+        except Article.DoesNotExist:
+            messages.error(request, f"Article with number {article_nbr} not found.")
+        except Exception as e:
+            messages.error(request, f"Unexpected error occurred, please retry later !")
+    return redirect(request.META.get('HTTP_REFERER', '/'))
 
 def remove_article_from_project(request, project_pk, article_pk):
     article = Article.objects.get(id=article_pk)
@@ -99,18 +108,18 @@ def remove_article_from_project(request, project_pk, article_pk):
     if request.method == "POST": 
         article.projects.remove(project)
         messages.success(request,'Article has been removed from project succesfully')
-    return redirect('project-detail', project.project_nbr)
+    return redirect('project-detail', project.id)
   
 @csrf_exempt
 def get_projectsByKeyWord(request):
     keyword = request.GET.get('keyword', '')
-    projects = Project.objects.filter(project_nbr__icontains=keyword).values_list('project_nbr', flat=True)[:20]
+    projects = Project.objects.filter(project_nbr__icontains=keyword).values('id', 'project_nbr')[:20]
     return JsonResponse(list(projects), safe=False)
 
 @csrf_exempt
 def get_clientsByKeyWord(request):
     keyword = request.GET.get('keyword', '')
-    projects = Client.objects.filter(client_nbr__icontains=keyword).values_list('client_nbr', flat=True)[:20]
+    projects = Client.objects.filter(client_nbr__icontains=keyword).values('id','client_nbr')[:20]
     return JsonResponse(list(projects), safe=False)
 
 def upload_file_to_project(request, project_pk):
@@ -141,15 +150,14 @@ def download_file(request, file_pk):
     return serve(request, project_file.file.name, document_root=settings.MEDIA_ROOT)
 
 def remove_file(request, file_pk):
-    try:
-        project_file = File.objects.get(id=file_pk)
-        if project_file:
-            project_file.delete()
-            path = f'media/{project_file.file}'
-            os.remove(path)
-    except:
-        messages.error(request,'Error occured !')
-    return redirect('project-detail',project_file.project.project_nbr)
+    if request.method == 'POST':
+        try:
+            project_file = File.objects.get(id=file_pk)
+            if project_file:
+                project_file.delete()
+        except:
+            messages.error(request,'Error occured !')
+        return redirect('project-detail',project_file.project.id)
        
 
 def client_detail(request, pk):
@@ -159,41 +167,31 @@ def client_detail(request, pk):
     context = {'client':client,'countries':countries,'page':page_name}
     return render(request, 'client.html', context)
 
-def client_create(request):
+def create_client(request):
     if request.method == 'POST':
         form = ClientForm(request.POST)
         if form.is_valid():
             form.save()
-            return JsonResponse({'message': 'Client has been added successfully'})
-        
-        for err in form.errors:
-            print(err)
-        return JsonResponse({'message':'An error occured ! please retry again'})
-    page_name = 'add-client'
-    nbr_clients = Client.objects.all().count()
-    client_nbr  = "C{0}".format(nbr_clients + 1)
+            return JsonResponse({'message':'Client has been created successfully'})
+        return JsonResponse({'message':get_message_error(form)})
     countries = Country.objects.all()
     languages = Language.objects.all()
     representatives = Representative.objects.all()
-    context = {'client_nbr':client_nbr,
-               'countries':countries,
+    page_name = 'add-client'
+    context = {'countries':countries,
                'languages':languages,
                'representatives':representatives,
                'page':page_name}
     return render(request, 'client.html', context)
 
-def client_edit(request, pk):
+def edit_client(request, pk):
     client = get_object_or_404(Client, pk=pk)
     if request.method == 'POST':
         form = ClientForm(request.POST, instance=client)
         if form.is_valid():
             form.save()
             return JsonResponse({'message': 'Client has been modified successfully'})
-        
-        for err in form.errors:
-            print(err)
-        return JsonResponse({'message':'An error occured ! please retry again'})
-    
+        return JsonResponse({'message': get_message_error(form)})
     countries = Country.objects.all()
     page_name = 'update-client'
     representatives = Representative.objects.all()
@@ -221,7 +219,7 @@ def get_clientByRef(request, ref):
                'page':page_name}
     return render(request, 'client.html', context)
 
-def client_delete(request, pk):
+def delete_client(request, pk):
     if request.method == "POST":
         client = get_object_or_404(Client, pk=pk)
         client.delete()
@@ -326,31 +324,21 @@ def get_supplierDetailByKeyWord(request):
 @csrf_exempt
 def get_buyersByKeyWord(request):
     keyword = request.GET.get('keyword', '')
-    buyers = Buyer.objects.filter(name__icontains=keyword)[:5]
-    buyer_data = []
-    if len(buyers):
-        for buyer in buyers:
-            buyer_data.append({
-                'id':buyer.id,
-                'name': buyer.name ,
-                'email': buyer.email or '',
-                'phone_number':buyer.phone_number or '',
-            })
-    # Return the buyers' data as JSON response
-    return JsonResponse({'buyers': buyer_data})
+    buyers = Buyer.objects.filter(name__icontains=keyword).values('id','name','email','phone_number')[:5]
+    return JsonResponse(list(buyers), safe=False)
 
 def create_buyer(request): 
     if request.method == 'POST':
         form = BuyerForm(request.POST)
         if form.is_valid():
             buyer = form.save()
+            print(buyer)
             buyer = {
                 'id': buyer.id,
                 'name': buyer.name,
             }
             return JsonResponse({'buyer': buyer})
-        else:
-            return JsonResponse({'error': 'error occured !'})
+        return JsonResponse({'error': get_message_error(form)})
     return render(request, 'buyer.html')
 
 def update_buyer(request, pk):
@@ -364,12 +352,9 @@ def update_buyer(request, pk):
                 'name': buyer.name,
             }
             return JsonResponse({'buyer': buyer})
-        return JsonResponse({'message': 'error occured !'})
+        return JsonResponse({'error': get_message_error(form)})
     context = {'buyer': buyer}
-    return render(request, 'buyer_edit.html', context)
-
-def chose_buyer(request):
-    return render(request, 'chose_buyer.html')
+    return render(request, 'buyer.html', context)
